--- a/src/ReverseProxy/Routing/HeaderMatcher.cs
+++ b/src/ReverseProxy/Routing/HeaderMatcher.cs
@@ -3,6 +3,8 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
+using Microsoft.Net.Http.Headers;
 using Yarp.ReverseProxy.Configuration;
 
 namespace Yarp.ReverseProxy.Routing;
@@ -30,11 +32,18 @@ internal sealed class HeaderMatcher
         {
             throw new ArgumentException($"Header values must not be specified when using '{nameof(HeaderMatchMode.Exists)}'.", nameof(values));
         }
+        if (values is not null && values.Any(string.IsNullOrEmpty))
+        {
+            throw new ArgumentNullException(nameof(values), "Header values must be not be empty.");
+        }
 
         Name = name;
-        Values = values ?? Array.Empty<string>();
+        Values = values?.ToArray() ?? Array.Empty<string>();
         Mode = mode;
-        IsCaseSensitive = isCaseSensitive;
+        Comparison = isCaseSensitive ? StringComparison.Ordinal : StringComparison.OrdinalIgnoreCase;
+
+        // Should this extend to Set-Cookie?
+        Separator = name.Equals(HeaderNames.Cookie, StringComparison.OrdinalIgnoreCase) ? ';' : ',';
     }
 
     /// <summary>
@@ -46,7 +55,7 @@ internal sealed class HeaderMatcher
     /// Returns a read-only collection of acceptable header values used during routing.
     /// At least one value is required unless <see cref="Mode"/> is set to <see cref="HeaderMatchMode.Exists"/>.
     /// </summary>
-    public IReadOnlyList<string> Values { get; }
+    public string[] Values { get; }
 
     /// <summary>
     /// Specifies how header values should be compared (e.g. exact matches Vs. by prefix).
@@ -54,11 +63,7 @@ internal sealed class HeaderMatcher
     /// </summary>
     public HeaderMatchMode Mode { get; }
 
-    /// <summary>
-    /// Specifies whether header value comparisons should ignore case.
-    /// When <c>true</c>, <see cref="StringComparison.Ordinal" /> is used.
-    /// When <c>false</c>, <see cref="StringComparison.OrdinalIgnoreCase" /> is used.
-    /// Defaults to <c>false</c>.
-    /// </summary>
-    public bool IsCaseSensitive { get; }
+    public StringComparison Comparison { get; }
+
+    public char Separator { get; }
 }