--- a/src/ReverseProxy/Routing/HeaderMetadata.cs
+++ b/src/ReverseProxy/Routing/HeaderMetadata.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace Yarp.ReverseProxy.Routing;
 
@@ -13,9 +14,9 @@ internal sealed class HeaderMetadata : IHeaderMetadata
 {
     public HeaderMetadata(IReadOnlyList<HeaderMatcher> matchers)
     {
-        Matchers = matchers ?? throw new ArgumentNullException(nameof(matchers));
+        Matchers = matchers?.ToArray() ?? throw new ArgumentNullException(nameof(matchers));
     }
 
     /// <inheritdoc/>
-    public IReadOnlyList<HeaderMatcher> Matchers { get; }
+    public HeaderMatcher[] Matchers { get; }
 }