--- a/src/ReverseProxy/Routing/IQueryParameterMetadata.cs
+++ b/src/ReverseProxy/Routing/IQueryParameterMetadata.cs
@@ -13,5 +13,5 @@ internal interface IQueryParameterMetadata
     /// <summary>
     /// One or more matchers to apply to the request query parameters.
     /// </summary>
-    IReadOnlyList<QueryParameterMatcher> Matchers { get; }
+    QueryParameterMatcher[] Matchers { get; }
 }