--- a/src/ReverseProxy/Routing/QueryParameterMatcher.cs
+++ b/src/ReverseProxy/Routing/QueryParameterMatcher.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using Yarp.ReverseProxy.Configuration;
 
 namespace Yarp.ReverseProxy.Routing;
@@ -30,11 +31,15 @@ internal sealed class QueryParameterMatcher
         {
             throw new ArgumentException($"Query parameter values must not be specified when using '{nameof(QueryParameterMatchMode.Exists)}'.", nameof(values));
         }
+        if (values is not null && values.Any(string.IsNullOrEmpty))
+        {
+            throw new ArgumentNullException(nameof(values), "Query parameter values must not be empty.");
+        }
 
         Name = name;
-        Values = values ?? Array.Empty<string>();
+        Values = values?.ToArray() ?? Array.Empty<string>();
         Mode = mode;
-        IsCaseSensitive = isCaseSensitive;
+        Comparison = isCaseSensitive ? StringComparison.Ordinal : StringComparison.OrdinalIgnoreCase;
     }
 
     /// <summary>
@@ -46,7 +51,7 @@ internal sealed class QueryParameterMatcher
     /// Returns a read-only collection of acceptable query parameter values used during routing.
     /// At least one value is required unless <see cref="Mode"/> is set to <see cref="QueryParameterMatchMode.Exists"/>.
     /// </summary>
-    public IReadOnlyList<string> Values { get; }
+    public string[] Values { get; }
 
     /// <summary>
     /// Specifies how query parameter values should be compared (e.g. exact matches Vs. contains).
@@ -54,11 +59,5 @@ internal sealed class QueryParameterMatcher
     /// </summary>
     public QueryParameterMatchMode Mode { get; }
 
-    /// <summary>
-    /// Specifies whether query parameter value comparisons should ignore case.
-    /// When <c>true</c>, <see cref="StringComparison.Ordinal" /> is used.
-    /// When <c>false</c>, <see cref="StringComparison.OrdinalIgnoreCase" /> is used.
-    /// Defaults to <c>false</c>.
-    /// </summary>
-    public bool IsCaseSensitive { get; }
+    public StringComparison Comparison { get; }
 }