Fixes #1494

Matching now treats multiple values as *OR*s. I.e. `Exact`/`Prefix` must match *any* of the request values.
`NotContains` means the value must *not match any* request header.

Values are further split based on the separator `,` (or `;` for `Cookie`). A single outer pair of quotes is also removed before matching.

This PR is generally slightly faster on simple (single-value) inputs, and (as expected) slower on multi-value ones.
Modes that may observe a perf degradation are `Exact` and `Prefix`.
I don't think the difference is prohibitive to the point that having an opt-out switch is needed.

Benchmark scenarios are defined [here](https://gist.github.com/MihaZupan/ccac2200ff7ba2f509060ee1a061e230#file-headerroutingbenchmark-cs-L30-L64).

| Method |         Scenario |      Mean |    Error |   StdDev | Ratio |
|------- |----------------- |----------:|---------:|---------:|------:|
|    Old |   SimpleContains | 292.78 ns | 0.957 ns | 0.848 ns |  1.00 |
|    New |   SimpleContains | 243.82 ns | 0.823 ns | 0.687 ns |  0.83 |
|        |                  |           |          |          |       |
|    Old |      SimpleExact | 158.98 ns | 0.529 ns | 0.469 ns |  1.00 |
|    New |      SimpleExact | 117.40 ns | 0.236 ns | 0.184 ns |  0.74 |
|        |                  |           |          |          |       |
|    Old |     SimplePrefix | 104.09 ns | 0.284 ns | 0.222 ns |  1.00 |
|    New |     SimplePrefix |  97.69 ns | 0.512 ns | 0.454 ns |  0.94 |
|        |                  |           |          |          |       |
|    Old | MultipleContains | 294.57 ns | 1.679 ns | 1.402 ns |  1.00 |
|    New | MultipleContains | 291.51 ns | 3.079 ns | 2.880 ns |  0.99 |
|        |                  |           |          |          |       |
|    Old |    MultipleExact | 124.28 ns | 0.237 ns | 0.185 ns |  1.00 |
|    New |    MultipleExact | 175.34 ns | 0.626 ns | 0.523 ns |  1.41 |
|        |                  |           |          |          |       |
|    Old |   MultiplePrefix | 130.89 ns | 1.529 ns | 1.431 ns |  1.00 |
|    New |   MultiplePrefix | 176.16 ns | 0.513 ns | 0.428 ns |  1.35 |

ToDo: Updating docs with the new behavior